+++ src/ska_low_mccs_spshw/station/tests/test_tile_beamformer.py
@@ -68,6 +68,7 @@ class TestTileBeamformer(BaseDaqTest):
         tile_trls: list[str],
         subrack_trls: list[str],
         daq_trl: str,
+        wren_trl: str,
     ) -> None:
         """
         Initialise a new instance.
@@ -76,6 +77,7 @@ class TestTileBeamformer(BaseDaqTest):
         :param tile_trls: trls of tiles the station has.
         :param subrack_trls: trls of subracks the station has.
         :param daq_trl: trl of the daq the station has.
+        :param wren_trl: trl of the wren the station has.
         :param component_manager: SpsStation component manager under test.
         """
         # Random seed for repeatability
@@ -90,7 +92,9 @@ class TestTileBeamformer(BaseDaqTest):
         self._ref_pol = randomiser.randrange(0, TileData.POLS_PER_ANTENNA, 1)
 
         self._start_freq = 156.25e6  # Hz
-        super().__init__(component_manager, logger, tile_trls, subrack_trls, daq_trl)
+        super().__init__(
+            component_manager, logger, tile_trls, subrack_trls, daq_trl, wren_trl
+        )
 
     def _send_beam_data(self: TestTileBeamformer) -> None:
         """Send beam data to the DAQ."""