+++ tests/unit/station/conftest.py
@@ -585,6 +585,7 @@ def station_self_check_manager_fixture(
         tile_trls=tile_trls,
         subrack_trls=subrack_trls,
         daq_trl="",
+        wren_trl="",
     )
     # Jank to get around https://github.com/python/mypy/issues/3115 and
     # https://github.com/python/mypy/issues/16509
@@ -595,6 +596,7 @@ def station_self_check_manager_fixture(
             tile_trls=list(tile_trls),
             subrack_trls=list(subrack_trls),
             daq_trl="",
+            wren_trl="",
         )
         for tpm_test in [
             PassTest,
@@ -608,6 +610,7 @@ def station_self_check_manager_fixture(
             tile_trls=list(tile_trls),
             subrack_trls=list(subrack_trls),
             daq_trl="",
+            wren_trl="",
         )
         for tpm_test in [
             ErrorTest,