+++ tests/unit/station/conftest.py
@@ -198,6 +198,9 @@ def mock_daq_device_proxy_fixture() -> MockDeviceBuilder:
     builder.add_result_command(
         "Start", result_code=ResultCode.QUEUED, status="Task queued"
     )
+    builder.add_result_command(
+        "Configure", result_code=ResultCode.OK, status="Configure completed OK."
+    )
     return builder()