+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 14.0.2
 
 * [SKB-1445] Update ``AcquireDataForCalibration`` to be more resilient to DAQ errors.
+* [THORN-125] Improve observability when UNPROGRAMMED.
+* [THORN-125] Bump daq version: 6.1.0 -> 6.1.1
 
 ## 14.0.1