+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Version History
 
-## Unreleased
+## 14.0.2
 
+* [SKB-1445] Update ``AcquireDataForCalibration`` to be more resilient to DAQ errors.
 * [THORN-125] Improve observability when UNPROGRAMMED.
 * [THORN-125] Bump daq version: 6.1.0 -> 6.1.1
 * [THORN-611] Add `WrenTRL` device property to SpsStation for WREN TRL